SCC363-33 Create Registration Flow
Biiiig commit.
Created some DTOs (Data transport objects) which are meant to be used for transporting objects between the client and server. This didnt work but but might be applicable to other tasks.

When going to /register, the users input is validated before being added to the database. The UI needs to represent any errors.

Made edits to allow you to access the h2-console and moved to using the database for storing users.